refactor(glossary): 용어 표준 3계층 정합 — 컬럼명·타입·길이 표준화 (V83~V85)
공공/산업 표준용어사전 전수 감사 결과 LS_* 비정합분을 표준에 정합.
런타임 동작 무변경(명명·타입 표준화), 관제 MNG_* 무변경, 전체 1743 테스트 GREEN + ddl-validate 부팅 OK.

Changes:
- V83: 컬럼 rename 5 (KPST_*→DE_IDNTF_*, REGISTERED_AT→REG_AT, REQ_KIND_CD→REQ_KND_CD, POLL_ATTEMPT_CNT→POLL_ATMPT_CNT)
- V84: 타입정합 (FRM_NO/VDO_FRM_NO INTEGER→BIGINT, truncation 길이확대, REQ/RES_PAYLOAD→TEXT) + 데이터마트 뷰 3종 재생성
- V85: 여부 CHAR(1) 통일 13건(공공 여부C1) + YesNoConverter(Boolean↔'Y'/'N')
- 엔티티/서비스 ~30: @Column·@JdbcTypeCode(CHAR)·필드타입(Integer→Long, Math.toIntExact 경계) 정합

+++ backend/src/main/java/kr/co/cudo/authoring/batch/entity/LsDataSrc.java
@@ -39,11 +39,11 @@ public class LsDataSrc {
     private Long rawSn;
 
     @Column(name = "FRM_NO", nullable = false)
-    private Integer frameNo;
+    private Long frameNo;
 
     /** 실제 영상 내 디코더 0-base 프레임 위치. FRM_NO(추출순번)와 의미 구분 — 재비식별 재추출용. nullable. */
     @Column(name = "VDO_FRM_NO", nullable = true)
-    private Integer videoFrameNo;
+    private Long videoFrameNo;
 
     @Column(name = "SRC_FILE_PATH_NM", nullable = false, length = 500)
     private String srcFilePathNm;
@@ -61,7 +61,7 @@ public class LsDataSrc {
     private LocalDateTime updDt;
 
     @Builder
-    private LsDataSrc(Long rawSn, Integer frameNo, Integer videoFrameNo, String srcFilePathNm, LocalDateTime shtDt) {
+    private LsDataSrc(Long rawSn, Long frameNo, Long videoFrameNo, String srcFilePathNm, LocalDateTime shtDt) {
         this.rawSn = rawSn;
         this.frameNo = frameNo;
         this.videoFrameNo = videoFrameNo;
@@ -71,7 +71,7 @@ public class LsDataSrc {
     }
 
     /** 원본(RAW) 프레임 row 생성 (videoFrameNo 미지정 = null). */
-    public static LsDataSrc create(Long rawSn, int frameNo, String srcFilePathNm, LocalDateTime shtDt) {
+    public static LsDataSrc create(Long rawSn, long frameNo, String srcFilePathNm, LocalDateTime shtDt) {
         return LsDataSrc.builder()
                 .rawSn(rawSn)
                 .frameNo(frameNo)
@@ -81,7 +81,7 @@ public class LsDataSrc {
     }
 
     /** 원본(RAW) 프레임 row 생성 (실제 영상 프레임 위치 videoFrameNo 보존). */
-    public static LsDataSrc create(Long rawSn, int frameNo, Integer videoFrameNo, String srcFilePathNm, LocalDateTime shtDt) {
+    public static LsDataSrc create(Long rawSn, long frameNo, Long videoFrameNo, String srcFilePathNm, LocalDateTime shtDt) {
         return LsDataSrc.builder()
                 .rawSn(rawSn)
                 .frameNo(frameNo)
